Better camera/positioning code
Allow me to move around & see things from different angles.

The bearing computations are not working; currently, 0° points down -z,
whereas my expectation was that it pointed down +x. z makes sense, given
that by default, z is into/out of the screen, but -z is currently
opposite what I expect, and I'd like to fix that first, I think. So,
while the code is terrible, it is usable.