Move all glyph rendering to a single spot
This greatly reduces the number of spots we're calling into FreeType.
(All of which involve `unsafe`.) In particular, we reduce the number of
callback functions to one: so if that function can do the panic handling
and span capture correctly, then the rest is normal safe rust.

We also make that function do nothing but capture the spans from
FreeType. (And to the stack, if possible.)